[ADD] write support (#37)
* Improve lzf error handling.

* Unify package descriptions.

* Add support for superblock extension.

* Add support for HSDS point selection.

* Feature/static constructor and struct records (#30)

* Migrate Level 1 files

* Migrate BTree1 and parts of BTree2

* Finish BTree2 migrations

* Migrate symbol table node

* Migrate symbol table entry

* Migrate local heap

* Migrate global heap

* Migrate global heap block for virtual datasets

* Migrate fractal heap ids

* Migrate rest of fractal heap

* Migrate free space manager

* Migrate shared object header message table

* Migrate object header

* Migrate nil, dataspace and link info messages

* Migrate more messages

* Migrate the rest - not yet tested

* Fix tests

---------


* Prepare update of references implementation and speed up tests

* Fix incorrect data layout message v1/v2 decoding method

* Add support for region references

* Finish object / region references implementation

* Implement simple variable-length data reading algorithm

* Clean up

* Clean up

* Finish variable-length datasets and improve tests

* Intermediate commit

* Finish filter improvements

* Clean up

* Simplify filters

* Remove unnecessary interfaces

* Prepare first writer test

* Begin writing the object header

* Intermediate commit.

* First successfully written file (groups only)

* Finish test.

* Do not create same object multiple times

* Write attributes

* Dump works

* Make tests work

* Fix pipeline

* Fix pipeline

* Try to fix pipeline

* Print path

* Make h5dump executable

* Test

* Maybe it just works?

* Use absolute path

* Fix pwd -.-

* Update samples

* Improve tests

* Make enum support generic

* Fix wrong type size.

* Intermediate commit

* Work on type size determination.

* Intermediate commit.

* Finish untested generic data type message creation algorithm

* Prepare data conversions

* Fixes

* Intermediate commit

* Intermediate commit

* Clean up

* Attribute write tests work except for variable length data

* Improve attribute writer

* Restructure tests

* Add support for Memory<T>

* Add support for strings (only h5dump still complains)

* Make simple string attribute work

* All attribute write tests except reference type are working

* Fix enum not being serialized correctly

* Improve test

* Add null value support for variable length data

* Remove duplicate test data

* Skip intermediate buffer

* Add support for multi dimensional arrays

* Temporarily disable HSDS tests.

* Test new IsReferenceOrContainsReferences implementation

* Remove useless conversion

* Add half and decimal precision

* Remove decimal test

* Disable HSDS tests

* Workaround for failing tests (run writing tests separately)

* Add Uint128 and Int128 tests

* Add support for Half and U/Int128 types to read API

* Clean up

* Edit README.md

* Add support for fixed length strings

* Improve array forming

* Make global heap flushable and global heap collection size configurable

* Prepare usage of free space manager for object header

* Intermediate commit

* Intermediate commit

* Intermediate commit

* Implement GetEncodeSize();

* Fix some bugs

* Fix more errors

* Clean up

* Intermediate commit

* Merge writing types into main code base

* Fixes

* Fixes

* Clean up

* Prepare compact dataset support

* Fix wrong root group address

* Fix remaining bugs

* Clean up

* Add support for compact and contiguous datasets

* Implicit index nearly works ...

* Add fail message

* Fix tests

* Prepare proper chunking

* Intermediate commit

* Clean up

* Convert all array like types to memory

* Intermediate commit

* Intermediate commit

* Repair attribute tests

* Add explanation why tuple names cannot be extracted

* Use EncodeDelegate everywhere

* Prepare general structure to make use of write chunk cache

* Prepare reuse of H5D classes

* Right before reading chunked data

* Write uncompressed chunked datasets

* Prepare fixed array index usage

* Rename Serialize to Write

* Extend writing API to allow passing streams

* Intermediate commit.

* Some fixes

* Prepare filter pipeline

* Improve filter registration process

* Enable passing options to filters

* Prepare usage of fixed array index

* Prepare writing of fixed array chunk index

* Encode the fixed array header and data block

* Make filtered chunks work

* Simplify specifing the filter pipline.

* Remove filter id enums and replace them by constants

* Rename chunk dimensions to chunks

* Add mass tests

* Simplify complex group test

* Rename group tests to link tests

* Update ShuffleAvx2 implementation

* Add shuffle tests

* Fix ShuffleAvx2 implementation

* Fix invalid windows path character

* Use specific commit

* Add fletcher32 filter support

* Prepare single chunk index implementation

* Repair filter tests

* Auto-chunk if filter pipeline is configured

* Remove Layout.Address property

* Add BZip2 write support

* Add Blosc2 compression support

* Add LZF filter support

* Clean up

* Clean up

* Add ISA-L inflate support

* Do not implement scale-offset compressor - too much effort

* Adapt exception

* Prepare reenabling of HSDS tests

* Fix HSDS tests

* Prepare writing of data after file layout phase

* Intermediate commit

* Nearly able to write data after file layout phase

* Make deferred writing work for each layout/index except single chunk

* Still not working

* Intermediate

* Deferred write to single chunk indexed data is working now

* Prepare memory & file selections

* Add first write with selection tests

* Finish selection support

* Finish write support (fill value support comes later)

* Prepare docfx

* Prepare docfx

---------

+++ src/PureHDF/VOL/Native/FileFormat/Level2/Enums/MessageFlags.cs
@@ -1,7 +1,7 @@
 ﻿namespace PureHDF.VOL.Native;
 
 [Flags]
-internal enum MessageFlags : ushort
+internal enum MessageFlags : byte
 {
     NoFlags = 0,
     Constant = 1,
